+++ server/src/main/java/com/main/project/restaurant/dto/RestaurantResponseDto.java
@@ -1,0 +1,2 @@
+package com.main.project.restaurant.dto;public class RestaurantResponseDto {
+}
