+++ server/src/main/java/com/main/project/restaurant/mapper/RestaurantMapper.java
@@ -1,0 +1,2 @@
+package com.main.project.restaurant.mapper;public interface RestaurantMapper {
+}
